Dont load hibernate instrumentation unless correct version of hibernate is on the classpath
