+++ dd-java-agent/instrumentation/hibernate/core-3.5/src/main/java/datadog/trace/instrumentation/hibernate/core/v3_5/SessionFactoryInstrumentation.java
@@ -27,11 +27,7 @@ import org.hibernate.Session;
 import org.hibernate.StatelessSession;
 
 @AutoService(Instrumenter.class)
-public class SessionFactoryInstrumentation extends Instrumenter.Default {
-
-  public SessionFactoryInstrumentation() {
-    super("hibernate", "hibernate-core");
-  }
+public class SessionFactoryInstrumentation extends AbstractHibernateInstrumentation {
 
   @Override
   public Map<String, String> contextStore() {
@@ -40,18 +36,6 @@ public class SessionFactoryInstrumentation extends Instrumenter.Default {
     stores.put("org.hibernate.StatelessSession", SessionState.class.getName());
     stores.put("org.hibernate.SharedSessionContract", SessionState.class.getName());
     return stores;
-  }
-
-  @Override
-  public String[] helperClassNames() {
-    return new String[] {
-      "datadog.trace.instrumentation.hibernate.SessionState",
-      "datadog.trace.agent.decorator.BaseDecorator",
-      "datadog.trace.agent.decorator.ClientDecorator",
-      "datadog.trace.agent.decorator.DatabaseClientDecorator",
-      "datadog.trace.agent.decorator.OrmClientDecorator",
-      "datadog.trace.instrumentation.hibernate.HibernateDecorator",
-    };
   }
 
   @Override
